// src/constants/types.ts
export type colorsType =
  | "r000"
  | "softPrimaryColor"
  | "vividPrimaryColor"
  | "f300"
  | "f200"
  | "f100"
  | "f000"
  | "g300"
  | "g200"
  | "g100"
  | "g000";

export type sizesType = "xs" | "sm" | "md" | "lg" | "xl";

export type variantsType = "primary" | "secondary";

export type typographyType =
  | "title1"
  | "heading1"
  | "heading2"
  | "heading3"
  | "body1"
  | "body2"
  | "body3"
  | "caption";